+++ src/domain/index.js
@@ -1,0 +1,3 @@
+export { default as Lotto } from './Lotto.js';
+export { default as LottoCompany } from './LottoCompany.js';
+export { default as LottoShop } from './LottoShop.js';
